+++ packages/block-library/src/breadcrumbs/index.php
@@ -17,23 +17,34 @@
  * @return string Returns the post breadcrumb for hierarchical post types.
  */
 function render_block_core_breadcrumbs( $attributes, $content, $block ) {
-	// Exclude breadcrumbs from special contexts.
-	if ( is_home() || is_front_page() ) {
+	$is_home_or_front_page = is_home() || is_front_page();
+	if ( ! $attributes['showOnHomePage'] && $is_home_or_front_page ) {
 		return '';
 	}
 
 	$breadcrumb_items = array();
 
 	if ( $attributes['showHomeLink'] ) {
-		$breadcrumb_items[] = sprintf(
-			'<a href="%s">%s</a>',
-			esc_url( home_url() ),
-			esc_html__( 'Home' )
-		);
+		// On home/front page, show as current page instead of link.
+		if ( $is_home_or_front_page ) {
+			$breadcrumb_items[] = sprintf(
+				'<span aria-current="page">%s</span>',
+				esc_html__( 'Home' )
+			);
+		} else {
+			$breadcrumb_items[] = sprintf(
+				'<a href="%s">%s</a>',
+				esc_url( home_url( '/' ) ),
+				esc_html__( 'Home' )
+			);
+		}
 	}
 
-	// Handle search results.
-	if ( is_search() ) {
+	// Handle home and front page.
+	if ( $is_home_or_front_page ) {
+		$breadcrumb_items = block_core_breadcrumbs_maybe_add_paged( $breadcrumb_items );
+	} elseif ( is_search() ) {
+		// Handle search results.
 		$search_query       = get_search_query();
 		$breadcrumb_items[] = sprintf(
 			'<span aria-current="page">%s</span>',
@@ -85,7 +96,16 @@ function render_block_core_breadcrumbs( $attributes, $content, $block ) {
 			$breadcrumb_items = array_merge( $breadcrumb_items, block_core_breadcrumbs_get_terms_breadcrumbs( $post_id, $post_type ) );
 		}
 		// Add current post title (not linked).
-		$breadcrumb_items[] = sprintf( '<span aria-current="page">%s</span>', get_the_title( $post ) );
+		$title = get_the_title( $post );
+		if ( strlen( $title ) === 0 ) {
+			$title = __( '(no title)' );
+		}
+		$breadcrumb_items[] = sprintf( '<span aria-current="page">%s</span>', $title );
+	}
+
+	// Remove last item if disabled.
+	if ( ! $attributes['showLastItem'] && ! empty( $breadcrumb_items ) ) {
+		array_pop( $breadcrumb_items );
 	}
 
 	if ( empty( $breadcrumb_items ) ) {
@@ -172,10 +192,14 @@ function block_core_breadcrumbs_get_hierarchical_post_type_breadcrumbs( $post_id
 	$ancestors        = array_reverse( $ancestors );
 
 	foreach ( $ancestors as $ancestor_id ) {
+		$title = get_the_title( $ancestor_id );
+		if ( strlen( $title ) === 0 ) {
+			$title = __( '(no title)' );
+		}
 		$breadcrumb_items[] = sprintf(
 			'<a href="%s">%s</a>',
 			esc_url( get_permalink( $ancestor_id ) ),
-			get_the_title( $ancestor_id )
+			$title
 		);
 	}
 	return $breadcrumb_items;