+++ packages/boot/src/store/selectors.ts
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import type { State } from './types';
+
+export function getMenuItems( state: State ) {
+	return Object.values( state.menuItems );
+}